Add piratey name theme to generator
Also, allow optional theme='' parameter to the namegenerator, primarily
so treasure hunts can specify only piratey names.
The 'pirate' theme is not linked to any biomes by default, so will not
currently appear unless specifically requested.  Possibly it could be
given a non-zero chance on beach and island biomes (25, 26, 14, 15, 16)
though it may not be compatible with the current entrance styles and so
is not enabled for now